Enable warnings as errors and burn down warnings (#51)
As many warnings as possible were enabled, and all compiler complaints
were addressed. Note that there is one major warning missing here:
-Woverloaded-virtual. This warning should be added in the future, but
requires some more involved architectural work on multiple models. It's
best saved for a few dedicated pull requests.

The C version of the double to words model was also deprecated. It's
unused and wasn't worth updating to fix the compiler warnings.

Refs #42

+++ models/utilities/enhanced_logging/src/enhanced_logging_variable.cc
@@ -242,7 +242,7 @@ EnhancedLogging_Variable::process_ref_attributes()
      variable name and the array is expanded beyond the limit of the "name"
      specification.
   */
-  int index_count = 0;
+  unsigned int index_count = 0;
   if (name.back() == ']') {
     std::string name_base = name;
     while (name_base.back() == ']') {
@@ -263,7 +263,7 @@ EnhancedLogging_Variable::process_ref_attributes()
   */
   for (unsigned int ii = index_count; ii < 8; ++ii) {
     if (ref->attr->index[ii].size == 0) { break;}
-    array_size.push_back( ref->attr->index[ii].size);
+    array_size.push_back( static_cast<unsigned int>(ref->attr->index[ii].size));
   }
   // Flag the variable as trick-processed so we only need to do this once.
   trick_processed = true;
